Move weight&Balance page to pages folder
