+++ AviationApp/AviationApp/AviationApp.csproj
@@ -19,7 +19,7 @@
     <Compile Update="AppShell.xaml.cs">
       <DependentUpon>AppShell.xaml</DependentUpon>
     </Compile>
-    <Compile Update="WeightAndBalance\WeightAndBalancePage.xaml.cs">
+    <Compile Update="Pages\WeightAndBalancePage.xaml.cs">
       <DependentUpon>WeightAndBalancePage.xaml</DependentUpon>
     </Compile>
   </ItemGroup>
@@ -52,7 +52,7 @@
     <EmbeddedResource Update="Pages\SettingsPage.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
-    <EmbeddedResource Update="WeightAndBalance\WeightAndBalancePage.xaml">
+    <EmbeddedResource Update="Pages\WeightAndBalancePage.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
   </ItemGroup>
