+++ AviationApp/AviationApp/Pages/WeightAndBalancePage.xaml.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 
-namespace AviationApp.WeightAndBalance
+namespace AviationApp.Pages
 {
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class WeightAndBalancePage : ContentPage
